Move types to FlowTypes file and add ContainerGeneratorPaths type
